+++ src/contexts/AppProvider.tsx
@@ -1,0 +1,22 @@
+import { createContext, useContext } from "react";
+import { User } from "../types/user";
+import { Notification } from "../types/notification";
+
+export interface AppContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+  notifications: Notification[];
+  addNotification: (message: string, type: Notification["type"]) => void;
+  removeNotification: (id: number) => void;
+}
+
+export const AppContext = createContext<AppContextType | undefined>(undefined);
+
+export const useAppContext = () => {
+  const context = useContext(AppContext);
+  if (context === undefined) {
+    throw new Error("useAppContext must be used within an AppProvider");
+  }
+  return context;
+};
