+++ src/contexts/theme/ThemeContext.ts
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+type ThemeContextType = {
+  theme: string;
+  toggleTheme: () => void;
+};
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+  undefined,
+);
